+++ predicators/ground_truth_models/__init__.py
@@ -7,12 +7,9 @@ import numpy as np
 from predicators import utils
 from predicators.envs import BaseEnv, get_or_create_env
 from predicators.envs.doors import DoorsEnv
-from predicators.envs.painting import PaintingEnv
 from predicators.envs.pddl_env import _PDDLEnv
 from predicators.envs.playroom import PlayroomEnv
-from predicators.envs.repeated_nextto_painting import RepeatedNextToPaintingEnv
 from predicators.envs.satellites import SatellitesEnv
-from predicators.envs.tools import ToolsEnv
 from predicators.envs.touch_point import TouchOpenEnv
 from predicators.settings import CFG
 from predicators.structs import NSRT, Array, GroundAtom, LiftedAtom, Object, \
@@ -135,12 +132,8 @@ utils.import_submodules(__path__, __name__)
 
 def deprecated_get_gt_nsrts(env_name: str) -> Set[NSRT]:
     """Create ground truth NSRTs for an env."""
-    if env_name in ("painting", "repeated_nextto_painting"):
-        nsrts = _get_painting_gt_nsrts(env_name)
-    elif env_name == "tools":
-        nsrts = _get_tools_gt_nsrts(env_name)
-    elif env_name in ("playroom", "playroom_simple", "playroom_hard",
-                      "playroom_simple_clear"):
+    if env_name in ("playroom", "playroom_simple", "playroom_hard",
+                    "playroom_simple_clear"):
         nsrts = _get_playroom_gt_nsrts(env_name)
     elif env_name in ("repeated_nextto", "repeated_nextto_ambiguous"):
         nsrts = _get_repeated_nextto_gt_nsrts(env_name)
@@ -198,779 +191,6 @@ def _get_options_by_names(env_name: str,
     options = get_gt_options(env_name)
     name_to_option = {o.name: o for o in options}
     return [name_to_option[name] for name in names]
-
-
-def _get_painting_gt_nsrts(env_name: str) -> Set[NSRT]:
-    """Create ground truth NSRTs for PaintingEnv."""
-    obj_type, box_type, lid_type, shelf_type, robot_type = \
-        _get_types_by_names(env_name, ["obj", "box", "lid", "shelf", "robot"])
-
-    (InBox, InShelf, IsBoxColor, IsShelfColor, GripperOpen, OnTable, \
-        NotOnTable, HoldingTop, HoldingSide, Holding, IsWet, IsDry, IsDirty, \
-        IsClean) = \
-        _get_predicates_by_names(
-            env_name, ["InBox", "InShelf", "IsBoxColor", "IsShelfColor",
-                        "GripperOpen", "OnTable", "NotOnTable", "HoldingTop",
-                        "HoldingSide", "Holding", "IsWet", "IsDry", "IsDirty",
-                        "IsClean"])
-
-    Pick, Wash, Dry, Paint, Place, OpenLid = _get_options_by_names(
-        env_name, ["Pick", "Wash", "Dry", "Paint", "Place", "OpenLid"])
-
-    if env_name == "repeated_nextto_painting":
-        (NextTo, NextToBox, NextToShelf, NextToTable) = \
-         _get_predicates_by_names(
-             env_name, ["NextTo", "NextToBox", "NextToShelf", "NextToTable"])
-        MoveToObj, MoveToBox, MoveToShelf = _get_options_by_names(
-            env_name, ["MoveToObj", "MoveToBox", "MoveToShelf"])
-
-    nsrts = set()
-
-    # PickFromTop
-    obj = Variable("?obj", obj_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, robot]
-    option_vars = [robot, obj]
-    option = Pick
-    preconditions = {
-        LiftedAtom(GripperOpen, [robot]),
-        LiftedAtom(OnTable, [obj])
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {LiftedAtom(Holding, [obj]), LiftedAtom(HoldingTop, [obj])}
-    delete_effects = {LiftedAtom(GripperOpen, [robot])}
-
-    def pickfromtop_sampler(state: State, goal: Set[GroundAtom],
-                            rng: np.random.Generator,
-                            objs: Sequence[Object]) -> Array:
-        del state, goal, rng, objs  # unused
-        return np.array([1.0], dtype=np.float32)
-
-    pickfromtop_nsrt = NSRT("PickFromTop", parameters, preconditions,
-                            add_effects, delete_effects, set(), option,
-                            option_vars, pickfromtop_sampler)
-    nsrts.add(pickfromtop_nsrt)
-
-    # PickFromSide
-    obj = Variable("?obj", obj_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, robot]
-    option_vars = [robot, obj]
-    option = Pick
-    preconditions = {
-        LiftedAtom(GripperOpen, [robot]),
-        LiftedAtom(OnTable, [obj])
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {LiftedAtom(Holding, [obj]), LiftedAtom(HoldingSide, [obj])}
-    delete_effects = {LiftedAtom(GripperOpen, [robot])}
-
-    def pickfromside_sampler(state: State, goal: Set[GroundAtom],
-                             rng: np.random.Generator,
-                             objs: Sequence[Object]) -> Array:
-        del state, goal, rng, objs  # unused
-        return np.array([0.0], dtype=np.float32)
-
-    pickfromside_nsrt = NSRT("PickFromSide", parameters, preconditions,
-                             add_effects, delete_effects, set(), option,
-                             option_vars, pickfromside_sampler)
-    nsrts.add(pickfromside_nsrt)
-
-    # Wash
-    obj = Variable("?obj", obj_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, robot]
-    option_vars = [robot]
-    option = Wash
-    preconditions = {
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(IsDry, [obj]),
-        LiftedAtom(IsDirty, [obj])
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {LiftedAtom(IsWet, [obj]), LiftedAtom(IsClean, [obj])}
-    delete_effects = {LiftedAtom(IsDry, [obj]), LiftedAtom(IsDirty, [obj])}
-
-    wash_nsrt = NSRT("Wash", parameters, preconditions, add_effects,
-                     delete_effects, set(), option, option_vars, null_sampler)
-    nsrts.add(wash_nsrt)
-
-    # Dry
-    obj = Variable("?obj", obj_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, robot]
-    option_vars = [robot]
-    option = Dry
-    preconditions = {
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(IsWet, [obj]),
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {LiftedAtom(IsDry, [obj])}
-    delete_effects = {LiftedAtom(IsWet, [obj])}
-
-    dry_nsrt = NSRT("Dry", parameters, preconditions, add_effects,
-                    delete_effects, set(), option, option_vars, null_sampler)
-    nsrts.add(dry_nsrt)
-
-    # PaintToBox
-    obj = Variable("?obj", obj_type)
-    box = Variable("?box", box_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, box, robot]
-    option_vars = [robot]
-    option = Paint
-    preconditions = {
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(IsDry, [obj]),
-        LiftedAtom(IsClean, [obj])
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {LiftedAtom(IsBoxColor, [obj, box])}
-    delete_effects = set()
-
-    def painttobox_sampler(state: State, goal: Set[GroundAtom],
-                           rng: np.random.Generator,
-                           objs: Sequence[Object]) -> Array:
-        del goal, rng  # unused
-        box_color = state.get(objs[1], "color")
-        return np.array([box_color], dtype=np.float32)
-
-    painttobox_nsrt = NSRT("PaintToBox", parameters, preconditions,
-                           add_effects, delete_effects, set(), option,
-                           option_vars, painttobox_sampler)
-    nsrts.add(painttobox_nsrt)
-
-    # PaintToShelf
-    obj = Variable("?obj", obj_type)
-    shelf = Variable("?shelf", shelf_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, shelf, robot]
-    option_vars = [robot]
-    option = Paint
-    preconditions = {
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(IsDry, [obj]),
-        LiftedAtom(IsClean, [obj])
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {LiftedAtom(IsShelfColor, [obj, shelf])}
-    delete_effects = set()
-
-    def painttoshelf_sampler(state: State, goal: Set[GroundAtom],
-                             rng: np.random.Generator,
-                             objs: Sequence[Object]) -> Array:
-        del goal, rng  # unused
-        shelf_color = state.get(objs[1], "color")
-        return np.array([shelf_color], dtype=np.float32)
-
-    painttoshelf_nsrt = NSRT("PaintToShelf", parameters, preconditions,
-                             add_effects, delete_effects, set(), option,
-                             option_vars, painttoshelf_sampler)
-    nsrts.add(painttoshelf_nsrt)
-
-    # PlaceInBox
-    obj = Variable("?obj", obj_type)
-    box = Variable("?box", box_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, box, robot]
-    option_vars = [robot]
-    option = Place
-    preconditions = {
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(HoldingTop, [obj]),
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextToBox, [robot, box]))
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {
-        LiftedAtom(InBox, [obj, box]),
-        LiftedAtom(GripperOpen, [robot]),
-        LiftedAtom(NotOnTable, [obj]),
-    }
-    delete_effects = {
-        LiftedAtom(HoldingTop, [obj]),
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(OnTable, [obj]),
-    }
-    if env_name == "repeated_nextto_painting":
-        # (Not)OnTable is affected by moving, not placing, in rnt_painting.
-        # So we remove it from the add and delete effects here.
-        add_effects.remove(LiftedAtom(NotOnTable, [obj]))
-        delete_effects.remove(LiftedAtom(OnTable, [obj]))
-
-    def placeinbox_sampler(state: State, goal: Set[GroundAtom],
-                           rng: np.random.Generator,
-                           objs: Sequence[Object]) -> Array:
-        del goal  # unused
-        x = state.get(objs[0], "pose_x")
-        if env_name == "painting":
-            y = rng.uniform(PaintingEnv.box_lb, PaintingEnv.box_ub)
-            z = state.get(objs[0], "pose_z")
-        elif env_name == "repeated_nextto_painting":
-            y = rng.uniform(RepeatedNextToPaintingEnv.box_lb,
-                            RepeatedNextToPaintingEnv.box_ub)
-            z = RepeatedNextToPaintingEnv.obj_z
-        return np.array([x, y, z], dtype=np.float32)
-
-    placeinbox_nsrt = NSRT("PlaceInBox", parameters, preconditions,
-                           add_effects, delete_effects, set(), option,
-                           option_vars, placeinbox_sampler)
-    nsrts.add(placeinbox_nsrt)
-
-    # PlaceInShelf
-    obj = Variable("?obj", obj_type)
-    shelf = Variable("?shelf", shelf_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, shelf, robot]
-    option_vars = [robot]
-    option = Place
-    preconditions = {
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(HoldingSide, [obj]),
-    }
-    if env_name == "repeated_nextto_painting":
-        preconditions.add(LiftedAtom(NextToShelf, [robot, shelf]))
-        preconditions.add(LiftedAtom(NextTo, [robot, obj]))
-    add_effects = {
-        LiftedAtom(InShelf, [obj, shelf]),
-        LiftedAtom(GripperOpen, [robot]),
-        LiftedAtom(NotOnTable, [obj]),
-    }
-    delete_effects = {
-        LiftedAtom(HoldingSide, [obj]),
-        LiftedAtom(Holding, [obj]),
-        LiftedAtom(OnTable, [obj]),
-    }
-    if env_name == "repeated_nextto_painting":
-        # (Not)OnTable is affected by moving, not placing, in rnt_painting.
-        # So we remove it from the add and delete effects here.
-        add_effects.remove(LiftedAtom(NotOnTable, [obj]))
-        delete_effects.remove(LiftedAtom(OnTable, [obj]))
-
-    def placeinshelf_sampler(state: State, goal: Set[GroundAtom],
-                             rng: np.random.Generator,
-                             objs: Sequence[Object]) -> Array:
-        del goal  # unused
-        x = state.get(objs[0], "pose_x")
-        if env_name == "painting":
-            y = rng.uniform(PaintingEnv.shelf_lb, PaintingEnv.shelf_ub)
-            z = state.get(objs[0], "pose_z")
-        elif env_name == "repeated_nextto_painting":
-            y = rng.uniform(RepeatedNextToPaintingEnv.shelf_lb,
-                            RepeatedNextToPaintingEnv.shelf_ub)
-            z = RepeatedNextToPaintingEnv.obj_z
-        return np.array([x, y, z], dtype=np.float32)
-
-    placeinshelf_nsrt = NSRT("PlaceInShelf", parameters, preconditions,
-                             add_effects, delete_effects, set(), option,
-                             option_vars, placeinshelf_sampler)
-    nsrts.add(placeinshelf_nsrt)
-
-    # OpenLid
-    lid = Variable("?lid", lid_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [lid, robot]
-    option_vars = [robot, lid]
-    option = OpenLid
-    preconditions = {LiftedAtom(GripperOpen, [robot])}
-    add_effects = set()
-    delete_effects = set()
-
-    openlid_nsrt = NSRT("OpenLid",
-                        parameters, preconditions, add_effects, delete_effects,
-                        set(), option, option_vars, null_sampler)
-    nsrts.add(openlid_nsrt)
-
-    # PlaceOnTable
-    obj = Variable("?obj", obj_type)
-    robot = Variable("?robot", robot_type)
-    parameters = [obj, robot]
-    option_vars = [robot]
-    option = Place
-    if env_name == "painting":
-        # The environment is a little weird: the object is technically
-        # already OnTable when we go to place it on the table, because
-        # of how the classifier is implemented.
-        preconditions = {
-            LiftedAtom(Holding, [obj]),
-            LiftedAtom(OnTable, [obj]),
-        }
-        add_effects = {
-            LiftedAtom(GripperOpen, [robot]),
-        }
-        delete_effects = {
-            LiftedAtom(Holding, [obj]),
-            LiftedAtom(HoldingTop, [obj]),
-            LiftedAtom(HoldingSide, [obj]),
-        }
-    elif env_name == "repeated_nextto_painting":
-        preconditions = {
-            LiftedAtom(Holding, [obj]),
-            LiftedAtom(NextTo, [robot, obj]),
-            LiftedAtom(NextToTable, [robot]),
-        }
-        add_effects = {
-            LiftedAtom(GripperOpen, [robot]),
-            LiftedAtom(OnTable, [obj]),
-        }
-        delete_effects = {
-            LiftedAtom(Holding, [obj]),
-            LiftedAtom(HoldingTop, [obj]),
-            LiftedAtom(HoldingSide, [obj]),
-            LiftedAtom(NotOnTable, [obj]),
-        }
-
-    def placeontable_sampler(state: State, goal: Set[GroundAtom],
-                             rng: np.random.Generator,
-                             objs: Sequence[Object]) -> Array:
-        del goal  # unused
-        x = state.get(objs[0], "pose_x")
-        if env_name == "painting":
-            # Always release the object where it is, to avoid the
-            # possibility of collisions with other objects.
-            y = state.get(objs[0], "pose_y")
-            z = state.get(objs[0], "pose_z")
-        elif env_name == "repeated_nextto_painting":
-            # Release the object at a randomly-chosen position on the table
-            # such that it is NextTo the robot.
-            robot_y = state.get(objs[1], "pose_y")
-            table_lb = RepeatedNextToPaintingEnv.table_lb
-            table_ub = RepeatedNextToPaintingEnv.table_ub
-            y = state.get(objs[0], "pose_y")
-            z = state.get(objs[0], "pose_z")
-            if table_lb < robot_y < table_ub:
-                nextto_thresh = RepeatedNextToPaintingEnv.nextto_thresh
-                y_sample_lb = max(table_lb, robot_y - nextto_thresh)
-                y_sample_ub = min(table_ub, robot_y + nextto_thresh)
-                y = rng.uniform(y_sample_lb, y_sample_ub)
-                z = RepeatedNextToPaintingEnv.obj_z
-        return np.array([x, y, z], dtype=np.float32)
-
-    placeontable_nsrt = NSRT("PlaceOnTable", parameters, preconditions,
-                             add_effects, delete_effects, set(), option,
-                             option_vars, placeontable_sampler)
-    nsrts.add(placeontable_nsrt)
-
-    if env_name == "repeated_nextto_painting":
-
-        def moveto_sampler(state: State, goal: Set[GroundAtom],
-                           _rng: np.random.Generator,
-                           objs: Sequence[Object]) -> Array:
-            del goal  # unused
-            y = state.get(objs[1], "pose_y")
-            return np.array([y], dtype=np.float32)
-
-        # MoveToObj
-        robot = Variable("?robot", robot_type)
-        targetobj = Variable("?targetobj", obj_type)
-        parameters = [robot, targetobj]
-        option_vars = [robot, targetobj]
-        option = MoveToObj
-        preconditions = {
-            LiftedAtom(GripperOpen, [robot]),
-            LiftedAtom(OnTable, [targetobj]),
-        }
-        add_effects = {
-            LiftedAtom(NextTo, [robot, targetobj]),
-            LiftedAtom(NextToTable, [robot])
-        }
-        delete_effects = set()
-        # Moving could have us end up NextTo other objects, and
-        # can turn off being next to the box or the shelf.
-        ignore_effects = {NextTo, NextToBox, NextToShelf}
-
-        movetoobj_nsrt = NSRT("MoveToObj", parameters, preconditions,
-                              add_effects, delete_effects, ignore_effects,
-                              option, option_vars, moveto_sampler)
-        nsrts.add(movetoobj_nsrt)
-
-        # MoveToBox
-        robot = Variable("?robot", robot_type)
-        targetbox = Variable("?targetbox", box_type)
-        obj = Variable("?obj", obj_type)
-        parameters = [robot, targetbox, obj]
-        option_vars = [robot, targetbox]
-        option = MoveToBox
-        preconditions = {
-            LiftedAtom(NextTo, [robot, obj]),
-            LiftedAtom(Holding, [obj]),
-        }
-        add_effects = {
-            LiftedAtom(NextToBox, [robot, targetbox]),
-            LiftedAtom(NextTo, [robot, obj]),
-            LiftedAtom(NotOnTable, [obj])
-        }
-        delete_effects = {
-            LiftedAtom(NextToTable, [robot]),
-            LiftedAtom(OnTable, [obj])
-        }
-        # Moving could have us end up NextTo other objects.
-        ignore_effects = {NextTo}
-        movetobox_nsrt = NSRT("MoveToBox", parameters, preconditions,
-                              add_effects, delete_effects, ignore_effects,
-                              option, option_vars, moveto_sampler)
-        nsrts.add(movetobox_nsrt)
-
-        # MoveToShelf
-        robot = Variable("?robot", robot_type)
-        targetshelf = Variable("?targetshelf", shelf_type)
-        obj = Variable("?obj", obj_type)
-        parameters = [robot, targetshelf, obj]
-        option_vars = [robot, targetshelf]
-        option = MoveToShelf
-        preconditions = {
-            LiftedAtom(NextTo, [robot, obj]),
-            LiftedAtom(Holding, [obj]),
-        }
-        add_effects = {
-            LiftedAtom(NextToShelf, [robot, targetshelf]),
-            LiftedAtom(NextTo, [robot, obj]),
-            LiftedAtom(NotOnTable, [obj])
-        }
-        delete_effects = {
-            LiftedAtom(NextToTable, [robot]),
-            LiftedAtom(OnTable, [obj])
-        }
-        # Moving could have us end up NextTo other objects.
-        ignore_effects = {NextTo}
-        movetoshelf_nsrt = NSRT("MoveToShelf", parameters, preconditions,
-                                add_effects, delete_effects, ignore_effects,
-                                option, option_vars, moveto_sampler)
-        nsrts.add(movetoshelf_nsrt)
-
-    return nsrts
-
-
-def _get_tools_gt_nsrts(env_name: str) -> Set[NSRT]:
-    """Create ground truth NSRTs for ToolsEnv."""
-    robot_type, screw_type, screwdriver_type, nail_type, hammer_type, \
-        bolt_type, wrench_type, contraption_type = _get_types_by_names(
-            env_name, ["robot", "screw", "screwdriver", "nail", "hammer",
-                      "bolt", "wrench", "contraption"])
-
-    HandEmpty, HoldingScrew, HoldingScrewdriver, HoldingNail, HoldingHammer, \
-        HoldingBolt, HoldingWrench, ScrewPlaced, NailPlaced, BoltPlaced, \
-        ScrewFastened, NailFastened, BoltFastened, ScrewdriverGraspable, \
-        HammerGraspable = _get_predicates_by_names(
-            env_name, ["HandEmpty", "HoldingScrew", "HoldingScrewdriver",
-                      "HoldingNail", "HoldingHammer", "HoldingBolt",
-                      "HoldingWrench", "ScrewPlaced", "NailPlaced",
-                      "BoltPlaced", "ScrewFastened", "NailFastened",
-                      "BoltFastened", "ScrewdriverGraspable",
-                      "HammerGraspable"])
-
-    PickScrew, PickScrewdriver, PickNail, PickHammer, PickBolt, PickWrench, \
-        Place, FastenScrewWithScrewdriver, FastenScrewByHand, \
-        FastenNailWithHammer, FastenBoltWithWrench = _get_options_by_names(
-            env_name, ["PickScrew", "PickScrewdriver", "PickNail", "PickHammer",
-                      "PickBolt", "PickWrench", "Place",
-                      "FastenScrewWithScrewdriver", "FastenScrewByHand",
-                      "FastenNailWithHammer", "FastenBoltWithWrench"])
-
-    def placeback_sampler(state: State, goal: Set[GroundAtom],
-                          rng: np.random.Generator,
-                          objs: Sequence[Object]) -> Array:
-        # Sampler for placing an item back in its initial spot.
-        del goal, rng  # unused
-        _, item = objs
-        pose_x = state.get(item, "pose_x")
-        pose_y = state.get(item, "pose_y")
-        return np.array([pose_x, pose_y], dtype=np.float32)
-
-    def placeoncontraption_sampler(state: State, goal: Set[GroundAtom],
-                                   rng: np.random.Generator,
-                                   objs: Sequence[Object]) -> Array:
-        # Sampler for placing an item on a contraption.
-        del goal  # unused
-        _, _, contraption = objs
-        pose_lx = state.get(contraption, "pose_lx")
-        pose_ly = state.get(contraption, "pose_ly")
-        pose_ux = pose_lx + ToolsEnv.contraption_size
-        pose_uy = pose_ly + ToolsEnv.contraption_size
-        # Note: Here we just use the average (plus noise), to make sampler
-        # learning easier. We found that it's harder to learn to imitate the
-        # more preferable sampler, which uses rng.uniform over the bounds.
-        pose_x = pose_lx + (pose_ux - pose_lx) / 2.0 + rng.uniform() * 0.01
-        pose_y = pose_ly + (pose_uy - pose_ly) / 2.0 + rng.uniform() * 0.01
-        return np.array([pose_x, pose_y], dtype=np.float32)
-
-    nsrts = set()
-
-    # PickScrew
-    robot = Variable("?robot", robot_type)
-    screw = Variable("?screw", screw_type)
-    parameters = [robot, screw]
-    option_vars = [robot, screw]
-    option = PickScrew
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    add_effects = {LiftedAtom(HoldingScrew, [screw])}
-    delete_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    nsrts.add(
-        NSRT("PickScrew", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # PickScrewdriver
-    robot = Variable("?robot", robot_type)
-    screwdriver = Variable("?screwdriver", screwdriver_type)
-    parameters = [robot, screwdriver]
-    option_vars = [robot, screwdriver]
-    option = PickScrewdriver
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(ScrewdriverGraspable, [screwdriver])
-    }
-    add_effects = {LiftedAtom(HoldingScrewdriver, [screwdriver])}
-    delete_effects = {LiftedAtom(HandEmpty, [robot])}
-    nsrts.add(
-        NSRT("PickScrewDriver", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # PickNail
-    robot = Variable("?robot", robot_type)
-    nail = Variable("?nail", nail_type)
-    parameters = [robot, nail]
-    option_vars = [robot, nail]
-    option = PickNail
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    add_effects = {LiftedAtom(HoldingNail, [nail])}
-    delete_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    nsrts.add(
-        NSRT("PickNail", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # PickHammer
-    robot = Variable("?robot", robot_type)
-    hammer = Variable("?hammer", hammer_type)
-    parameters = [robot, hammer]
-    option_vars = [robot, hammer]
-    option = PickHammer
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(HammerGraspable, [hammer])
-    }
-    add_effects = {LiftedAtom(HoldingHammer, [hammer])}
-    delete_effects = {LiftedAtom(HandEmpty, [robot])}
-    nsrts.add(
-        NSRT("PickHammer", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # PickBolt
-    robot = Variable("?robot", robot_type)
-    bolt = Variable("?bolt", bolt_type)
-    parameters = [robot, bolt]
-    option_vars = [robot, bolt]
-    option = PickBolt
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    add_effects = {LiftedAtom(HoldingBolt, [bolt])}
-    delete_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-    }
-    nsrts.add(
-        NSRT("PickBolt", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # PickWrench
-    robot = Variable("?robot", robot_type)
-    wrench = Variable("?wrench", wrench_type)
-    parameters = [robot, wrench]
-    option_vars = [robot, wrench]
-    option = PickWrench
-    preconditions = {LiftedAtom(HandEmpty, [robot])}
-    add_effects = {LiftedAtom(HoldingWrench, [wrench])}
-    delete_effects = {LiftedAtom(HandEmpty, [robot])}
-    nsrts.add(
-        NSRT("PickWrench", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # PlaceScrewdriverBack
-    robot = Variable("?robot", robot_type)
-    screwdriver = Variable("?screwdriver", screwdriver_type)
-    parameters = [robot, screwdriver]
-    option_vars = [robot]
-    option = Place
-    preconditions = {
-        LiftedAtom(HoldingScrewdriver, [screwdriver]),
-        LiftedAtom(ScrewdriverGraspable, [screwdriver])
-    }
-    add_effects = {LiftedAtom(HandEmpty, [robot])}
-    delete_effects = {LiftedAtom(HoldingScrewdriver, [screwdriver])}
-    nsrts.add(
-        NSRT("PlaceScrewdriverBack", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, placeback_sampler))
-
-    # PlaceHammerBack
-    robot = Variable("?robot", robot_type)
-    hammer = Variable("?hammer", hammer_type)
-    parameters = [robot, hammer]
-    option_vars = [robot]
-    option = Place
-    preconditions = {
-        LiftedAtom(HoldingHammer, [hammer]),
-        LiftedAtom(HammerGraspable, [hammer])
-    }
-    add_effects = {LiftedAtom(HandEmpty, [robot])}
-    delete_effects = {LiftedAtom(HoldingHammer, [hammer])}
-    nsrts.add(
-        NSRT("PlaceHammerBack", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, placeback_sampler))
-
-    # PlaceWrenchBack
-    robot = Variable("?robot", robot_type)
-    wrench = Variable("?wrench", wrench_type)
-    parameters = [robot, wrench]
-    option_vars = [robot]
-    option = Place
-    preconditions = {LiftedAtom(HoldingWrench, [wrench])}
-    add_effects = {LiftedAtom(HandEmpty, [robot])}
-    delete_effects = {LiftedAtom(HoldingWrench, [wrench])}
-    nsrts.add(
-        NSRT("PlaceWrenchBack", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, placeback_sampler))
-
-    # PlaceScrewOnContraption
-    robot = Variable("?robot", robot_type)
-    screw = Variable("?screw", screw_type)
-    contraption = Variable("?contraption", contraption_type)
-    parameters = [robot, screw, contraption]
-    option_vars = [robot]
-    option = Place
-    preconditions = {LiftedAtom(HoldingScrew, [screw])}
-    add_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(ScrewPlaced, [screw, contraption])
-    }
-    delete_effects = {LiftedAtom(HoldingScrew, [screw])}
-    nsrts.add(
-        NSRT("PlaceScrewOnContraption", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars,
-             placeoncontraption_sampler))
-
-    # PlaceNailOnContraption
-    robot = Variable("?robot", robot_type)
-    nail = Variable("?nail", nail_type)
-    contraption = Variable("?contraption", contraption_type)
-    parameters = [robot, nail, contraption]
-    option_vars = [robot]
-    option = Place
-    preconditions = {LiftedAtom(HoldingNail, [nail])}
-    add_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(NailPlaced, [nail, contraption])
-    }
-    delete_effects = {LiftedAtom(HoldingNail, [nail])}
-    nsrts.add(
-        NSRT("PlaceNailOnContraption", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars,
-             placeoncontraption_sampler))
-
-    # PlaceBoltOnContraption
-    robot = Variable("?robot", robot_type)
-    bolt = Variable("?bolt", bolt_type)
-    contraption = Variable("?contraption", contraption_type)
-    parameters = [robot, bolt, contraption]
-    option_vars = [robot]
-    option = Place
-    preconditions = {LiftedAtom(HoldingBolt, [bolt])}
-    add_effects = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(BoltPlaced, [bolt, contraption])
-    }
-    delete_effects = {LiftedAtom(HoldingBolt, [bolt])}
-    nsrts.add(
-        NSRT("PlaceBoltOnContraption", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars,
-             placeoncontraption_sampler))
-
-    # FastenScrewWithScrewdriver
-    robot = Variable("?robot", robot_type)
-    screw = Variable("?screw", screw_type)
-    screwdriver = Variable("?screwdriver", screwdriver_type)
-    contraption = Variable("?contraption", contraption_type)
-    parameters = [robot, screw, screwdriver, contraption]
-    option_vars = [robot, screw, screwdriver, contraption]
-    option = FastenScrewWithScrewdriver
-    preconditions = {
-        LiftedAtom(HoldingScrewdriver, [screwdriver]),
-        LiftedAtom(ScrewPlaced, [screw, contraption])
-    }
-    add_effects = {LiftedAtom(ScrewFastened, [screw])}
-    delete_effects = set()
-    nsrts.add(
-        NSRT("FastenScrewWithScrewdriver", parameters, preconditions,
-             add_effects, delete_effects, set(), option, option_vars,
-             null_sampler))
-
-    # FastenScrewByHand
-    robot = Variable("?robot", robot_type)
-    screw = Variable("?screw", screw_type)
-    contraption = Variable("?contraption", contraption_type)
-    parameters = [robot, screw, contraption]
-    option_vars = [robot, screw, contraption]
-    option = FastenScrewByHand
-    preconditions = {
-        LiftedAtom(HandEmpty, [robot]),
-        LiftedAtom(ScrewPlaced, [screw, contraption])
-    }
-    add_effects = {LiftedAtom(ScrewFastened, [screw])}
-    delete_effects = set()
-    nsrts.add(
-        NSRT("FastenScrewByHand", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # FastenNailWithHammer
-    robot = Variable("?robot", robot_type)
-    nail = Variable("?nail", nail_type)
-    hammer = Variable("?hammer", hammer_type)
-    contraption = Variable("?contraption", contraption_type)
-    parameters = [robot, nail, hammer, contraption]
-    option_vars = [robot, nail, hammer, contraption]
-    option = FastenNailWithHammer
-    preconditions = {
-        LiftedAtom(HoldingHammer, [hammer]),
-        LiftedAtom(NailPlaced, [nail, contraption])
-    }
-    add_effects = {LiftedAtom(NailFastened, [nail])}
-    delete_effects = set()
-    nsrts.add(
-        NSRT("FastenNailWithHammer", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    # FastenBoltWithWrench
-    robot = Variable("?robot", robot_type)
-    bolt = Variable("?bolt", bolt_type)
-    wrench = Variable("?wrench", wrench_type)
-    contraption = Variable("?contraption", contraption_type)
-    parameters = [robot, bolt, wrench, contraption]
-    option_vars = [robot, bolt, wrench, contraption]
-    option = FastenBoltWithWrench
-    preconditions = {
-        LiftedAtom(HoldingWrench, [wrench]),
-        LiftedAtom(BoltPlaced, [bolt, contraption])
-    }
-    add_effects = {LiftedAtom(BoltFastened, [bolt])}
-    delete_effects = set()
-    nsrts.add(
-        NSRT("FastenBoltWithWrench", parameters, preconditions, add_effects,
-             delete_effects, set(), option, option_vars, null_sampler))
-
-    return nsrts
 
 
 def _get_playroom_gt_nsrts(env_name: str) -> Set[NSRT]:
