+++ predicators/ground_truth_models/tools/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for tools environment and variants."""
+
+from .nsrts import ToolsGroundTruthNSRTFactory
+
+__all__ = ["ToolsGroundTruthNSRTFactory"]
